requirements.txt
# ---------- global options ----------
--pre                 # ytdlpのnightly用
--extra-index-url https://pypi.org/simple
# --no-cache-dir


# Core dependencies
discord.py[voice]>=2.3.2
PyYAML>=6.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0

# Audio processing
yt-dlp
pycryptodomex

# Utilities
colorlog>=6.7.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.1